+++ src/targeted/targets/_init-targets.mjs
@@ -18,6 +18,7 @@ import { WhenInCombatTarget } from './conditional/when-in-combat-target.mjs';
 import { WhileAdjacentToTarget } from './conditional/while-adjacent-to-target.mjs';
 import { WhileEquippedTarget } from './conditional/while-equipped-target.mjs';
 import { WhileSharingSquareWithTarget } from './conditional/while-sharing-square-with-target.mjs';
+import { WhileWeaponGroupEquippedTarget } from './conditional/while-weapon-group-equipped-target.mjs';
 import { DamageTypeTarget } from "./damage-type-target.mjs";
 import { FinesseTarget } from './finesse-target.mjs';
 import { SelfTarget } from './self-target.mjs';
@@ -61,4 +62,5 @@ export const registerTargets = () => [
     WhileAdjacentToTarget,
     WhileEquippedTarget,
     WhileSharingSquareWithTarget,
+    WhileWeaponGroupEquippedTarget,
 ].forEach(Sources.registerSource);